Move FullAddressSearchInputProps to types.d
